fix clean architecture
