+++ api/coral/intrenal/handler/create_user.go
@@ -1,0 +1,33 @@
+package handler
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+)
+
+func (h *CoralHandler) CreateUserHandler(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
+		return
+	}
+
+	if r.Body == nil {
+		http.Error(w, "Empty request body", http.StatusBadRequest)
+		return
+	}
+
+	var user map[string]interface{}
+	if err := json.NewDecoder(r.Body).Decode(&user); err != nil {
+		http.Error(w, "Invalid JSON", http.StatusBadRequest)
+		return
+	}
+
+	if err := h.userUseCase.CreateUser(r.Context(), user); err != nil {
+		http.Error(w, fmt.Sprintf("Failed to create user in DynamoDB: %v", err), http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusCreated)
+	fmt.Fprintf(w, "User created: %v", user)
+}
