+++ api/coral/intrenal/handler/healthcheck.go
@@ -1,0 +1,15 @@
+package handler
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+func (h *CoralHandler) HealthCheckHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	response := map[string]string{"status": "ok"}
+	if err := json.NewEncoder(w).Encode(response); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+}
